small fix on type casting

+++ sql/now/_init.sql
@@ -65,7 +65,7 @@ OUTPUTS:
 DROP TABLE IF EXISTS _INIT_NOW_devdb;
 SELECT
 	distinct
-	(ogc_fid::integer + (SELECT MAX(uid) FROM _INIT_BIS_devdb)::text as uid,
+	(ogc_fid::integer + (SELECT MAX(uid::integer) FROM _INIT_BIS_devdb))::text as uid,
 	job_filing_number::text as job_number,
 	jobtype as job_type,
